+++ spec/features/dogma_effects_spec.rb
@@ -11,7 +11,7 @@ describe 'Get effects' do
 
   specify { expect(subject.scope).to eq(nil) }
 
-  specify { expect(subject.effects_ids.size).to eq(4123) }
+  specify { expect(subject.effect_ids.size).to eq(4123) }
 
-  specify { expect(subject.effects_ids.first).to eq(4) }
+  specify { expect(subject.effect_ids.first).to eq(4) }
 end